LP-2523 move helpers for courses to models
